receiving kitchener weather

+++ app/src/main/java/co/bagga/wetify/Models/Main.kt
@@ -9,5 +9,5 @@ class Main {
     var temp_min: Double? = null
     var temp_max: Double? = null
     var pressure: Double? = null
-    var humidity: Float? = null
+    var humidity: Double? = null
 }
